Move matchesAudience and stripPort tests to provisioner package.
